added types to component files
