+++ package.json
@@ -22,6 +22,7 @@
         "@types/react": "^18.3.16",
         "@types/react-dom": "^18.3.5",
         "@types/react-router-dom": "^5.3.3",
+        "@types/d3": "^7.4.3",
         "@vitejs/plugin-react": "^4.3.3",
         "autoprefixer": "^10.4.20",
         "eslint": "^9.13.0",
